+++ lib/widgets/screenshot.dart
@@ -1,8 +1,11 @@
+import 'dart:io' if (kIsWeb) 'dart:html';
 import 'dart:typed_data';
 
+import 'package:chart_builder/utils/toast/toast.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:image_gallery_saver/image_gallery_saver.dart';
 import 'package:share/share.dart';
+import 'package:path_provider/path_provider.dart';
 
 class ScreenshotDialog extends StatelessWidget {
   const ScreenshotDialog({
@@ -25,31 +28,61 @@ class ScreenshotDialog extends StatelessWidget {
   Widget build(BuildContext context) {
     return AlertDialog(
       title: Text('Screenshot taken'),
+      contentPadding: EdgeInsets.symmetric(vertical: 10, horizontal: 6),
       content: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
-          Card(child: Image.memory(bytes)),
+          Card(
+            child: Padding(
+              padding: EdgeInsets.all(4),
+              child: Image.memory(bytes),
+            ),
+          ),
           Text(
             'The ouside borders will NOT be saved',
             textAlign: TextAlign.center,
             style: Theme.of(context).textTheme.caption,
           ),
+          // TODO: make this work
+          Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              FlatButton(
+                child: Text('Save to device'),
+                color: Colors.amber,
+                onPressed: () async {
+                  final file = await getFile();
+                  await file.writeAsBytes(bytes);
+                  showTextToast(
+                    text: '$name was saved to\n${file.path}',
+                    context: context,
+                  );
+                },
+              ),
+              SizedBox(width: 4),
+              FlatButton(
+                child: Text('Share outside'),
+                color: Colors.yellow,
+                onPressed: () async {
+                  final file = await getFile();
+                  if (!file.existsSync())
+                    showTextToast(
+                      text: 'You must save the file to share it',
+                      context: context,
+                    );
+                },
+              ),
+            ],
+          ),
         ],
       ),
-      actions: [
-        FlatButton(
-          child: Text('Save to device'),
-          color: Colors.pink,
-          onPressed: () => ImageGallerySaver.saveImage(bytes, name: name),
-        ),
-        FlatButton(
-          child: Text('Share outside'),
-          color: Colors.indigo,
-          onPressed: () {
-            // TODO: share outside
-          },
-        ),
-      ],
     );
+  }
+
+  Future<File> getFile() async {
+    final directory = await getExternalStorageDirectory();
+    final name = this.name.trim().replaceAll(' ', '_');
+    final path = '${directory.path}/$name.png';
+    return File(path);
   }
 }