+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 Date format: DD/MM/YYYY
 
+## 1.0.2 - [15/02/2021]
+
+- **NEW**:
+  - **HIGHLIGHT**: It's now possible to save the pie chart to the device
+  - `AppTheme.isDark`
+- Long press in a section to edit a Pie chart, instead of tapping
+- Replace fab to `TextButton` in AppBar
+- Fixed chart preview in pie chart home
+
 ## 1.0.1 - [13/02/2021]
 
 - Settings Screen